--- app/containers/SignUpCompletePage/messages.js
@@ -1,7 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  title: '메일 전송 완료',
-  body: '아래 마이스누 웹메일로 인증메일을 전송하였습니다.<br />메일을 확인해 인증링크를 클릭해 주세요.',
-  back: '돌아가기',
-});
