+++ app/containers/SignUpPage/messages.js
@@ -23,4 +23,7 @@ export default defineMessages({
     text: '로그인하러 가기',
     hint: '계정이 이미 있다면?',
   },
+  completeTitle: '메일 전송 완료',
+  completeBody: '아래 마이스누 웹메일로 인증메일을 전송하였습니다.<br />메일을 확인해 인증링크를 클릭해 주세요.',
+  completeBack: '돌아가기',
 });
