+++ app/containers/SignUpPage/sagas.js
@@ -1,8 +1,7 @@
 import { take, call, put } from 'redux-saga/effects';
 import { request } from 'services/api';
 import { setAuthToken } from 'services/localStorage';
-import { Creators as GlobalActions } from 'global/reducer';
-import { browserHistory } from 'react-router';
+import { Creators as GlobalActions } from '../../global/reducer';
 import { Types, Creators as Actions } from './reducer';
 import { userInformation } from '../../global/sagas';
 
@@ -19,7 +18,6 @@ export function* signUp({ username, password, nickname, department_id }) {
     yield put(Actions.signUpSuccess());
     setAuthToken(response.data.meta.auth_token);
     yield call(userInformation);
-    browserHistory.push('sign_up/complete');
   } catch (error) {
     yield put(Actions.signUpFailure(error.errors));
   }
